+++ packages/Helm/src/utils/helm-grid-cell-renderer.ts
@@ -20,44 +20,28 @@ import {_package, getMonomerLib} from '../package';
 import {IMonomerLib} from '@datagrok-libraries/bio/src/types/index';
 import {GridCell, GridCellStyle, x} from 'datagrok-api/dg';
 
-class WrapLogger implements ILogger {
-  constructor(
-    private readonly base: ILogger
-  ) {}
-
-  error(message: any, params?: object | undefined, stackTrace?: string | undefined): void {
-    this.base.error(message, params, stackTrace);
-  }
-
-  warning(message: string, params?: object | undefined): void {
-    this.base.warning(message, params);
-  }
-
-  info(message: string, params?: object | undefined): void {
-    // this.base.info(message, params);
-  }
-
-  debug(message: string, params?: object | undefined): void {
-    // this.base.debug(message, params);
-  }
-}
-
 export class HelmGridCellRendererBack extends CellRendererBackAsyncBase<HelmProps, HelmAux> {
   private _auxList: (HelmAux | null)[];
 
   private monomerLib: IMonomerLib | null = null;
 
+  // eslint-disable-next-line max-len
+  private readonly uuid: string = wu.repeat(1).map(() => Math.floor((Math.random() * 36)).toString(36)).take(4).toArray().join('');
+
   constructor(
     gridCol: DG.GridColumn | null,
     tableCol: DG.Column<string>,
   ) {
-    super(gridCol, tableCol, new WrapLogger(_package.logger) /* _package.logger */, true);
+    super(gridCol, tableCol, _package.logger, true);
   }
 
   protected override reset(): void {
+    const logPrefix = `${this.toLog()}.reset()`;
+    this.logger.debug(`${logPrefix}, start`);
     super.reset();
     this._auxList = new Array<HelmAux | null>(this.tableCol.length).fill(null);
     if (this.gridCol && this.gridCol.dart) this.gridCol.grid?.invalidate();
+    this.logger.debug(`${logPrefix}, end`);
   }
 
   protected override getRenderService(): RenderServiceBase<HelmProps, HelmAux> {
@@ -71,6 +55,8 @@ export class HelmGridCellRendererBack extends CellRendererBackAsyncBase<HelmProp
   }
 
   protected override storeAux(gridCell: DG.GridCell, aux: HelmAux): void {
+    const logPrefix = `${this.toLog()}.storeAux()`;
+    this.logger.debug(`${logPrefix}, start`);
     if (gridCell.tableRowIndex !== null)
       this._auxList[gridCell.tableRowIndex] = aux;
   }
@@ -171,7 +157,8 @@ export class HelmGridCellRendererBack extends CellRendererBackAsyncBase<HelmProp
   // -- Handle events --
 
   private monomerLibOnChanged(_value: any): void {
-    this.reset();
+    this.dirty = true;
+    this.invalidateGrid();
   }
 
   static getOrCreate(gridCell: DG.GridCell): HelmGridCellRendererBack {